Pull request #3935: Cppcheck
Merge in SNORT/snort3 from ~RDEMPSTE/snort3:cppcheck to master

Squashed commit of the following:

commit e7663dd3be7fd74a91808f293de0222ea7a467ee
Author: Ron Dempster (rdempste) <rdempste@cisco.com>
Date:   Thu Oct 19 11:46:13 2023 -0400

    build: remove unused functions reported by cppcheck

commit ce623c51d7bb6a034d09b3700db74f1ccf229bc5
Author: Ron Dempster (rdempste) <rdempste@cisco.com>
Date:   Wed Jul 19 12:55:56 2023 -0400

    actions, detection, file_api, flow, stream: coverity fixes

commit bed4872d9259d16a345e61a15f766505c18b0c30
Author: Ron Dempster (rdempste) <rdempste@cisco.com>
Date:   Fri Jun 30 13:57:44 2023 -0400

    build: Address miscellaneous cppcheck warnings

+++ src/ips_options/ips_content.cc
@@ -322,10 +322,10 @@ static bool uniSearchReal(ContentData* cd, Cursor& c)
     else
     {
         // retry - adjust from start of last match
-        pos = c.get_pos() - cd->pmd.pattern_size + cd->match_delta;
+        pos = (int64_t)c.get_pos() - cd->pmd.pattern_size + cd->match_delta;
 
         if ( cd->depth_configured )
-            depth -= c.get_delta();
+            depth -= (int64_t)c.get_delta();
 
         if ( pos < 0 )
             return false;
@@ -361,6 +361,7 @@ static bool uniSearchReal(ContentData* cd, Cursor& c)
 
 static IpsOption::EvalStatus CheckANDPatternMatch(ContentData* idx, Cursor& c)
 {
+    // cppcheck-suppress unreadVariable
     RuleProfile profile(contentPerfStats);
 
     bool found = uniSearchReal(idx, c);
